requirements.txt
# LangGraph Agent Dependencies
# Week 1 Learning Project - LangGraph with Anthropic Claude

# LangGraph for agent workflows
langgraph>=0.2.0

# LangChain core and Anthropic integration
langchain>=0.3.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Development tools
pytest>=7.4.0
black>=23.12.0

# Note: API Key Configuration
# Set the following environment variable in your .env file:
# ANTHROPIC_API_KEY=your_anthropic_key_here